+++ lib/posix/Kconfig
@@ -55,3 +55,5 @@ source "lib/posix/Kconfig.signal"
 source "lib/posix/Kconfig.spinlock"
 source "lib/posix/Kconfig.timer"
 source "lib/posix/Kconfig.uname"
+
+rsource "shell/Kconfig"
