+++ lib/posix/shell/Kconfig
@@ -1,0 +1,8 @@
+# Copyright (c) 2024 Meta
+# SPDX-License-Identifier: Apache-2.0
+
+if SHELL
+
+rsource "Kconfig.uname"
+
+endif # SHELL
