+++ samples/posix/uname/prj.conf
@@ -1,4 +1,4 @@
 CONFIG_POSIX_API=y
 CONFIG_SHELL=y
-CONFIG_SHELL_GETOPT=y
+CONFIG_POSIX_UNAME_SHELL=y
 CONFIG_MP_MAX_NUM_CPUS=1
